% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_expression.R
\name{calculateRatios}
\alias{calculateRatios}
\title{Generate expression ratios for 2 given transcript ids for all samples.}
\usage{
calculateRatios(
  normCM,
  metadata,
  transcript_id1,
  transcript_id2,
  metadata_group,
  output_path
)
}
\arguments{
\item{normCM}{matrix that includes all normalized count matrices where
sample names are the columns and transcript ids are the rows. See data/Normalized_CM.rda for an example.}

\item{metadata}{matrix that includes the relevant group information for samples. See data/Experiment_info.rda for an example.}

\item{transcript_id1}{the id of the first transcript ID for comparing expression.}

\item{transcript_id2}{the id of the second transcript ID for comparing expression.}

\item{metadata_group}{the column name in metadata used to categorize samples into desired groups.}

\item{output_path}{desired path to save the output in a csv.}
}
\value{
Returns a matrix that includes the expression ratio of given transcript ids for each sample.
}
\description{
A function that returns a matrix that includes the expression ratio of given transcript ids for each sample in a given data.
}
\examples{
ratioDF <- calculateRatios(Normalized_CM, Experiment_info, "ENST00000316724.9", "ENST00000409400.1", "Biosample_summary")

}
\references{
Wickham H, François R, Henry L, Müller K, Vaughan D (2023). \emph{dplyr: A Grammar of Data Manipulation}. R
package version 1.1.3, \url{https://CRAN.R-project.org/package=dplyr}.
}
