% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineCM.R
\name{combineCounts}
\alias{combineCounts}
\title{Combine all kallisto count matrices to a single data frame.}
\usage{
combineCounts(countdir)
}
\arguments{
\item{countdir}{path to the directory with all the count matrices. This is the directory where you stored your kallisto outputs.
Note in order to use this package, kallisto is a required prerequisite step.}
}
\value{
Returns a matrix that combined every count matrices into one matrix.
}
\description{
A function that returns a matrix that includes all normalized count matrices where
sample names are the columns and transcript ids are the rows. See data/Normalized_CM.rda for an example.
}
\references{
R Core Team (2023). \emph{R: A Language and Environment for Statistical Computing}. R Foundation for Statistical
Computing, Vienna, Austria. \url{https://www.R-project.org/}.
}
