% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Experiment_info}
\alias{Experiment_info}
\title{Information of the samples provided for testing.}
\format{
A matrix with 4 columns
\describe{
\item{Accession}{Accession id for that sample}
\item{Biosample_age}{Age of the sample}
\item{Biosample_summary}{Description of the sample's condition}
\item{Quantification_file}{File id for that sample's transcript-level expression counts}
}
}
\source{
Barbara Wold Lab at California Institute of Technology, USA
}
\usage{
Experiment_info
}
\description{
This table is a subset of experiment_report.tsv downloaded from the ENCODE Project by clicking the 'Download TSV' button \href{https://www.encodeproject.org/report/?type=Experiment&status=released&internal_tags=RushAD&assay_title=total+RNA-seq&limit=200}{here}
}
\keyword{datasets}
