% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_expression.R
\name{generateBarPlot}
\alias{generateBarPlot}
\title{Generate a bar plot for the expression ratio created from calculateRatios.}
\usage{
generateBarPlot(data, x_label, y_label, title)
}
\arguments{
\item{data}{matrix generated from calculateRatios.}

\item{x_label}{A character vector specifying the desired label for the x-axis.}

\item{y_label}{A character vector specifying the desired label for the y-axis.}

\item{title}{A character vector specifying the desired title for the plot.}
}
\value{
Returns the bar plot.
}
\description{
A function that generates a bar plot for the expression ratio created from calculateRatios.
It is sorted by ratio from lowest to highest and also sorted by metadata_group
from the matrix generated from calculateRatios.
}
\examples{
ratioDF <- calculateRatios(Normalized_CM, Experiment_info, "ENST00000316724.9",
     "ENST00000409400.1", "Biosample_summary")
generateBarPlot(ratioDF, "Sample names", "Ratio", "Bar plot")

}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
