% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_expression.R
\name{generateBoxPlot}
\alias{generateBoxPlot}
\title{Generate a box plot for the expression ratio created from calculateRatios.}
\usage{
generateBoxPlot(
  data,
  x_label = NULL,
  y_label = NULL,
  title = NULL,
  order = NULL
)
}
\arguments{
\item{data}{matrix generated from calculateRatios.}

\item{x_label}{A character vector specifying the desired label for the x-axis. Default is NULL.}

\item{y_label}{A character vector specifying the desired label for the y-axis. Default is NULL.}

\item{title}{A character vector specifying the desired title for the plot. Default is NULL.}

\item{order}{desired ordered list of metadata_group names to sort the boxes in the plot. Default is NULL.}
}
\value{
Returns the box plot.
}
\description{
A function that generates a box plot for the expression ratio created from calculateRatios for each metadata_group.
}
\examples{
ratioDF <- calculateRatios(Normalized_CM, Experiment_info, "ENST00000316724.9",
    "ENST00000409400.1", "Biosample_summary")
generateBoxPlot(ratioDF, "Disease Type", "Ratio", "Box plot",
    c("Control", "mild cognitive impairment", "Cognitive impairment",
    "Alzheimer's disease, Cognitive impairment", "Alzheimer's disease"))

generateBoxPlot(ratioDF, "Disease Type", "Ratio", "Box plot")

}
\references{
Kassambara A (2023). \emph{ggpubr: 'ggplot2' Based Publication Ready Plots}. R package version 0.6.0,
\url{https://CRAN.R-project.org/package=ggpubr}.

Wickham H (2022). \emph{stringr: Simple, Consistent Wrappers for Common String Operations}. R package version
1.5.0, \url{https://CRAN.R-project.org/package=stringr}.
}
