% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Normalized_CM}
\alias{Normalized_CM}
\title{Matrix of all the normalized count matrices combined into one.}
\format{
A matrix with 24 columns
\describe{
\item{TranscriptIDs}{Transcript ID that can be searched on Ensembl}
\item{All other columns}{All other columns are the file names downloaded from the ENCODE Project}
}
}
\source{
Barbara Wold Lab at California Institute of Technology, USA
}
\usage{
Normalized_CM
}
\description{
This table is combines all the normalized expressions from files listed in Quantification_file in Experiment_info.
The count matrices were retrieved from the ENCODE Project where they were generated using kallisto.
Each row are the normalized transcript-level expression counts where each column are the sample file names.
}
\keyword{datasets}
