% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIsoformVisRNA.R
\name{runIsoformVisRNA}
\alias{runIsoformVisRNA}
\title{Launch Shiny App for IsoformVisRNA}
\usage{
runIsoformVisRNA()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for IsoformVisRNA.
The shiny app allows the user to generate bar and box plots through the UI.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

IsoformVisRNA::runIsoformVisRNA()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
